FuncLogRegModel.R
The FDA, or functional data analysis studies how to reconstruct the true functional form of the data, and creates a basis expansion to achieve it. For another hand, glm models can be applied in this way, to predict a binary response variable from a  functional predictor. A classic example about berkeley growth study is presented here to show the utility of functional analysis models.